Added histograms from matrices.
git-svn-id: http://svn.berlios.de/svnroot/repos/qtiplot/trunk@711 5a6a7de5-fb12-0410-b871-c33778c25c60

+++ qtiplot/src/VectorCurve.cpp
@@ -237,41 +237,38 @@ QwtDoubleRect VectorCurve::boundingRect() const
 QwtDoubleRect rect = QwtPlotCurve::boundingRect();
 QwtDoubleRect vrect = vectorEnd->boundingRect();
 
-if (d_style == XYXY)
-	{
-	rect.setTop(QMIN(rect.top(), vrect.top()));
-	rect.setBottom(QMAX(rect.bottom(), vrect.bottom()));
-	rect.setLeft(QMIN(rect.left(), vrect.left()));
-	rect.setRight(QMAX(rect.right(), vrect.right()));
-	}
-else
-	{
+if (d_style == XYXY){
+	rect.setTop(QMIN((double)rect.top(), (double)vrect.top()));
+	rect.setBottom(QMAX((double)rect.bottom(), (double)vrect.bottom()));
+	rect.setLeft(QMIN((double)rect.left(), (double)vrect.left()));
+	rect.setRight(QMAX((double)rect.right(), (double)vrect.right()));
+} else {
 	const double angle = vectorEnd->x(0);
 	double mag = vectorEnd->y(0);
 	switch(d_position)
 		{
 		case Tail:
-			rect.setTop(QMIN(rect.top(), rect.top()+mag*sin(angle)));
-			rect.setBottom(QMAX(rect.bottom(), rect.bottom()+mag*sin(angle)));
-			rect.setLeft(QMIN(rect.left(), rect.left()+mag*cos(angle)));
-			rect.setRight(QMAX(rect.right(), rect.right()+mag*cos(angle)));
+			rect.setTop(QMIN((double)rect.top(), (double)(rect.top()+mag*sin(angle))));
+			rect.setBottom(QMAX((double)rect.bottom(), (double)(rect.bottom()+mag*sin(angle))));
+			rect.setLeft(QMIN((double)rect.left(), (double)(rect.left()+mag*cos(angle))));
+			rect.setRight(QMAX((double)rect.right(), (double)(rect.right()+mag*cos(angle))));
 		break;
 
 		case Middle:
 			{
 			mag *= 0.5;
-			rect.setTop(QMIN(rect.top(), rect.top() - fabs(mag*sin(angle))));
-			rect.setBottom(QMAX(rect.bottom(), rect.bottom() + fabs(mag*sin(angle))));
-			rect.setLeft(QMIN(rect.left(), rect.left() - fabs(mag*cos(angle))));
-			rect.setRight(QMAX(rect.right(), rect.right() + fabs(mag*cos(angle))));
+			rect.setTop(QMIN((double)rect.top(), (double)(rect.top() - fabs(mag*sin(angle)))));
+			rect.setBottom(QMAX((double)rect.bottom(), (double)(rect.bottom() + fabs(mag*sin(angle)))));
+			rect.setLeft(QMIN((double)rect.left(), (double)(rect.left() - fabs(mag*cos(angle)))));
+			rect.setRight(QMAX((double)rect.right(), (double)(rect.right() + fabs(mag*cos(angle)))));
 			}
 		break;
 
 		case Head:
-			rect.setTop(QMIN(rect.top(), rect.top() - mag*sin(angle)));
-			rect.setBottom(QMAX(rect.bottom(), rect.bottom() - mag*sin(angle)));
-			rect.setLeft(QMIN(rect.left(), rect.left() - mag*cos(angle)));
-			rect.setRight(QMAX(rect.right(), rect.right() - mag*cos(angle)));
+			rect.setTop(QMIN((double)rect.top(), (double)(rect.top() - mag*sin(angle))));
+			rect.setBottom(QMAX((double)rect.bottom(), (double)(rect.bottom() - mag*sin(angle))));
+			rect.setLeft(QMIN((double)rect.left(), (double)(rect.left() - mag*cos(angle))));
+			rect.setRight(QMAX((double)rect.right(), (double)(rect.right() - mag*cos(angle))));
 		break;
 		}
 	}
@@ -280,8 +277,7 @@ return rect;
 
 void VectorCurve::updateColumnNames(const QString& oldName, const QString& newName, bool updateTableName)
 {
-    if (updateTableName)
-    {
+    if (updateTableName){
         QString s = title().text();
         QStringList lst = s.split("_", QString::SkipEmptyParts);
         if (lst[0] == oldName)
